Move RequestBuilder into clients package
